+++ test/bin/builder/inflater_builder_test.dart
@@ -3,12 +3,12 @@ import 'package:test/test.dart';
 import '../../../bin/src/builders/builder.dart';
 import '../../../bin/src/builders/inflaters.dart';
 
-
+/// Dart tests.
 void main() {
 
   test('Test InflaterBuilder dependency resolution', () async {
     final config = BuilderConfig();
-    await config.loadConfig("xwidget|res/default_config.yaml");
+    await config.loadConfig("xwidget_builder|res/default_config.yaml");
     await config.loadConfig("test/fixtures/res/xwidget_config.yaml");
 
     final builder = InflaterBuilder(config);
@@ -18,7 +18,7 @@ void main() {
 
   test('Test new inflater spec format', () async {
     final config = BuilderConfig();
-    await config.loadConfig("xwidget|res/default_config.yaml");
+    await config.loadConfig("xwidget_builder|res/default_config.yaml");
     await config.loadConfig("test/fixtures/res/xwidget_config.yaml");
 
     final builder = InflaterBuilder(config);
@@ -28,7 +28,7 @@ void main() {
 
   test('Test inflater spec generics', () async {
     final config = BuilderConfig();
-    await config.loadConfig("xwidget|res/default_config.yaml");
+    await config.loadConfig("xwidget_builder|res/default_config.yaml");
     await config.loadConfig("test/fixtures/res/xwidget_config.yaml");
 
     final builder = InflaterBuilder(config);