+++ test/bin/utils/path_resolver_test.dart
@@ -2,11 +2,15 @@ import 'package:test/test.dart';
 
 import '../../../bin/src/utils/path_resolver.dart';
 
-
+/// Dart tests.
 void main() {
-  test('Test package path resolution', () async {
-    PathResolver.packageRoot;
-    final path = await PathResolver.relativeToAbsolute("package:flutter/lib/test/image_data.dart");
-    expect(path, "");
+  test('Test package path resolution should throw exception with specific message', () async {
+    await expectLater(
+      PathResolver.relativeToAbsolute('package:flutter/src/widgets/fake_widget.dart'),
+      throwsA(predicate((e) {
+        print(e);
+        return e is Exception && e.toString().startsWith('Exception: Package');
+      }))
+    );
   });
 }