+++ test/fixtures/src/inflater_spec.dart
@@ -1,0 +1,27 @@
+import 'package:flutter/material.dart';
+
+// Define constants for all widgets and objects we want to generate inflaters for.
+// Used by the builder to generate inflaters.
+
+// ignore_for_file: unused_element
+
+const inflaters = [
+  AlertDialog,
+  AlwaysStoppedAnimation,
+  AlwaysStoppedAnimation<Color>,
+  AlwaysStoppedAnimation<int>,
+  AnimatedCrossFade,
+  AppBar,
+  BoxConstraints,
+  Center,
+  CircularProgressIndicator,
+  Column,
+  FloatingActionButton,
+  Icon,
+  MaterialApp,
+  Scaffold,
+  Text,
+  TextButton,
+  TextStyle,
+  ThemeData,
+];